+++ doc/src/changelog.md
@@ -6,9 +6,9 @@ title: Changelog
 # Pipe
 
 * refine design to get back pre-`0.4.x` symmetry (prevent the need for implementing both a summarizer and a resolver)
-* summarizer and resolver from node streams
 * ProtocolBuffer serialization
 * difference view indexed by key (allows to view updates in a different collection)
+* remove Promise-resolving in iterator-based APIs
 
 # On the road to 0.5.0